web_app/requirements.txt
Flask==2.3.3
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
requests==2.31.0
twilio==8.8.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Werkzeug==2.3.7

# WhatsApp Integration
pywhatkit==5.4
pyautogui>=0.9.54
pillow>=10.0.0
